requirements.txt
mysql-connector-python~=8.0.28
pywhatkit~=5.3
python-dotenv~=0.19.2